Revert

+++ tests/integ/bin/runtests.sh
@@ -2,7 +2,11 @@
 
 rm -rf testdata test_ticks.csv && mkdir -p testdata/mktsdb
 
-marketstore connect -d `pwd`/testdata/mktsdb <<- EOF >& /dev/null && diff -q bin/ticks-example-1-output.csv test_ticks.csv && echo "Passed" || echo "Test Failed"
+if [ $? -ne 0 ]; then
+  exit 1
+fi
+
+marketstore connect -d `pwd`/testdata/mktsdb <<- EOF
 \create TEST/1Min/TICK:Symbol/Timeframe/AttributeGroup Bid,Ask/float32 variable
 \getinfo TEST/1Min/TICK
 \load TEST/1Min/TICK bin/ticks-example-1.csv bin/ticks-example-1.yaml
@@ -10,15 +14,38 @@ marketstore connect -d `pwd`/testdata/mktsdb <<- EOF >& /dev/null && diff -q bin
 \show TEST/1Min/TICK 1970-01-01
 EOF
 
+if [ $? -ne 0 ]; then
+  exit 1
+fi
+
+diff -q bin/ticks-example-1-output.csv test_ticks.csv && echo "Passed"
+
+if [ $? -ne 0 ]; then
+  exit 1
+fi
+
 rm -f test_ticks.csv
 
-marketstore connect -d `pwd`/testdata/mktsdb <<- EOF >& /dev/null && diff -q bin/ticks-example-2-output.csv test_ticks.csv && echo "Passed" || echo "Test Failed"
-\create TEST/1Min/TICK:Symbol/Timeframe/AttributeGroup Bid,Ask/float32 variable
+if [ $? -ne 0 ]; then
+  exit 1
+fi
+
+marketstore connect -d `pwd`/testdata/mktsdb <<- EOF
 \getinfo TEST/1Min/TICK
 \load TEST/1Min/TICK bin/ticks-example-2.csv bin/ticks-example-2.yaml
 \o test_ticks.csv
 \show TEST/1Min/TICK 1970-01-01
 EOF
+
+if [ $? -ne 0 ]; then
+  exit 1
+fi
+
+diff -q bin/ticks-example-2-output.csv test_ticks.csv && echo "Passed"
+
+if [ $? -ne 0 ]; then
+  exit 1
+fi
 
 rm -rf testdata test_ticks.csv
